Issue #92: Prepared process details page to add ThreadComments feature.
