+++ bw_matchbox/assets/modules/ProcessDetailPage/@types/TInitParams.d.ts
@@ -1,0 +1,5 @@
+interface TInitParams {
+  threadComments: ThreadComments;
+  callbacks: TSharedHandlers;
+  sharedParams: TSharedParams;
+}
