+++ bw_matchbox/assets/modules/ProcessDetailPage/@types/TProcessDetailPageState.d.ts
@@ -1,0 +1,8 @@
+class ProcessDetailPageState {
+  isWaitlist: boolean;
+  setLoading: (isLoading: boolean) => void;
+  setWaitlist: (isWaitlist: boolean) => void;
+  setError: (error: string | Error | string[] | Error[]) => void;
+  clearError: () => void;
+}
+type TProcessDetailPageState = ProcessDetailPageState;
