+++ bw_matchbox/assets/modules/ProcessDetailPage/@types/TSharedParams.d.ts
@@ -1,0 +1,8 @@
+interface TSharedParams {
+  addAttributeUrl: string;
+  markMatchTypeUrl: string;
+  markMatchedUrl: string;
+  markAllMatchedUrl: string;
+  multimatch: string;
+  isWaitlist: boolean;
+}
