+++ bw_matchbox/assets/modules/ProcessDetailPage/@types/TWaitlistCommentDialogUserAction.d.ts
@@ -1,0 +1,4 @@
+interface TWaitlistCommentDialogUserAction {
+  comment: string;
+  status?: string;
+}
