+++ bw_matchbox/assets/modules/ProcessDetailPage/@types/components.d.ts
@@ -1,0 +1,2 @@
+// type TProcessDetailPageState = ProcessDetailPageState;
+type TProcessDetailPageHelpers = ProcessDetailPageHelpers;
